+++ MovieApp/Data/Entity/SearchInput.swift
@@ -6,67 +6,28 @@
 //
 
 import Foundation
-import Factory
 
-class MovieRemoteRepositoryImp: MovieRepository, APIClient {
-     var storage: StorageServicable
+struct SearchInput: Encodable {
+    var query: String
+    var isIncludeAdult: Bool
+    var language: String
+    var page: Int
     
-    init(storage: StorageServicable) {
-        self.storage = storage
+    enum CodingKeys: String, CodingKey {
+        case query
+        case isIncludeAdult = "include_adult"
+        case language
+        case page
     }
     
-    func search(query: String, page: Int) async throws -> PaginationDTO<MovieDTO> {
-        let input = SearchInput(query: query, page: page)
-        return try await request(endpoint: Router.search(input: input))
+    init(query: String, isIncludeAdult: Bool, language: String, page: Int) {
+        self.query = query
+        self.isIncludeAdult = isIncludeAdult
+        self.language = language
+        self.page = page
     }
     
-    func detail(id: String) async throws -> MovieDTO {
-        return try await request(endpoint: Router.detail(id: id))
-    }
-    
-}
-
-extension MovieRemoteRepositoryImp {
-    enum Router {
-        case search(input: SearchInput)
-        case detail(id: String)
-    }
-}
-
-extension MovieRemoteRepositoryImp.Router: Endpoint {
-    var path: String {
-        switch self {
-        case .search:
-            return API.search(.movie).url
-        case let .detail(id):
-            return  API.detail(.movie, id: id).url
-        }
-    }
-    
-    var method: HTTPMethod {
-        switch self {
-        case .search, .detail:
-            return .get
-        }
-    }
-    
-    var headers: [String : String]? {
-        return  [
-            "Content-Type"  : "application/json; charset=utf-8",
-            "accept"        : "application/json"
-        ]
-    }
-    
-    var queryItems: [URLQueryItem]? {
-        var query: [URLQueryItem] = [URLQueryItem(name: "api_key", value: Container.shared.storage().apiKey)]
-        switch self {
-        case let .search(input):
-            let items = input.queryItems ?? []
-            query.append(contentsOf: items)
-            return query
-            
-        default:
-            return query
-        }
+    init (query: String, page: Int) {
+        self.init(query: query, isIncludeAdult: false, language: "en-US", page: page)
     }
 }
